feat: restructure navigation components
